Extract sails object for jobs into Queue
